Added Task functionality

+++ src/main/java/com/epam/internship/exception/TaskNotFoundException.java
@@ -2,7 +2,7 @@ package com.epam.internship.exception;
 
 public class TaskNotFoundException extends  RuntimeException{
 
-    public TaskNotFoundException(Integer id) {
+    public TaskNotFoundException(Long id) {
         super("Task with id: " + id + " not found");
     }
 }
